chore: remove WHAT comments, extract isValidEmail to utils
L4: Remove ~15 section-label comments ({/* Toolbar */}, {/* Content */},
    etc.) from ImportVault, MigrationSuccess, MigrationHome — keep only
    WHY comments.

L5: Move hand-rolled email validation from VaultEmail into
    src/utils/isValidEmail.ts for reuse.
